backend/requirements.txt
flask==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
ultralytics==8.0.196
kafka-python==2.0.2
google-generativeai==0.3.1
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0